Crate psl_types

Common types for the public suffix implementation crates

The types in this crate assume that the input is valid UTF-8 encoded domain names. If input is potentially invalid, use a higher level crate like the addr crate.

Some implentations may also assume that the domain name is in lowercase and/or may only support looking up unicode domain names.

Structs

Domain

A registrable domain name

Info

Information about the suffix

Suffix

The suffix of a domain name

Enums

Type

Type of suffix

Traits

List

A list of all public suffixes